Estimator/requirements.txt
cloudpickle==1.6.0
urllib3==1.25.10
numpy==1.19.0
kafka-python==2.0.1